packages/backend/requirements.txt
fastapi==0.100.0
pydantic==1.10.8
uvicorn==0.22.0
pymongo==4.5.0
motor==3.3.1
python-jose==3.3.0
passlib==1.7.4
bcrypt==4.0.1
python-multipart==0.0.6
python-dotenv==1.0.0
pytest==7.4.0
httpx==0.24.1
pytest-cov==4.1.0
pytest-watch==4.2.0
openai==1.3.0
google-auth-oauthlib==1.0.0
google-api-python-client==2.97.0
python-dateutil==2.8.2 